+++ app/models/formatters/integer_formatter.rb
@@ -10,6 +10,7 @@ module Formatters
       elsif response.blank?
         response
       else
+        response = response.gsub(/^([-+]?)(0*)/, "\\1") if response.is_a?(String)
         Integer(format("%d", response))
       end
     rescue